requirements.txt
tensorflow==2.11.0
numpy==1.24.1
cvxpy==1.5.1
gym==0.16.0
tqdm==4.66.4
hydra-core==1.3.2
matplotlib==3.9.0
omegaconf==2.3.0
pyglet==1.5.0


